Fixed issue with showing sponsor images.  Also did a little refactoring of image size validation.
